change setExtensionUrls to setExtensionRuntimeData
